requirements.txt
fastapi
python-dotenv
httpx
uvicorn
pydantic
